requirements.txt
amqp==5.3.1
ansicon==1.89.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.1
blessed==1.21.0
Brotli==1.1.0
celery==5.3.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.8.0
Django==5.1.2
django-allauth==65.11.0
django-picklefield==3.3
django-q==1.3.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.59.0
html5lib==1.1
idna==3.10
jinxed==1.3.0
kombu==5.5.4
lxml==5.3.0
mysql-connector-python==9.1.0
oscrypto==1.3.0
packaging==25.0
pdfrw==0.4
pillow==11.0.0
prompt_toolkit==3.0.51
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.10.1
PyMySQL==1.1.1
pypdf==5.1.0
pyphen==0.17.2
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
redis==4.6.0
reportlab==4.2.5
requests==2.32.3
setuptools==80.9.0
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20250809
tzdata==2025.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
